aliases: extract zip to folder if necessary
Especially zips are sometimes packed without a folder inside to hold
everything. Extracting these archives leads to a mess on the current
directory.

This checks if there are more than one element in the root directory.
If so, create a folder based on the name of the zip, and extract into
that.