herbstluftwm: Add monitor setup
Setup all monitors and pad the main monitor for the panel

The external display of the notebook to the main monitor and if there
is no external display, the notebook display is being used as main
display.

The main monitor is the first display that is not LVDS1, or in case
there is no such display, LVDS1.